requirements.txt
numpy
cython
matplotlib
pytorch >= 0.4.0
rawpy
pytest
